+++ internal/cli/auth/auth.go
@@ -3,6 +3,7 @@ package auth
 import (
 	"context"
 	"fmt"
+	"net/url"
 
 	"github.com/spf13/cobra"
 
@@ -18,6 +19,7 @@ type AuthCmd struct {
 	Port                int
 	LocalBindAddress    string
 	RedirectURLHostname string
+	RedirectURL         string
 }
 
 func NewCommand(globalFlags *flags.GlobalFlags) *cobra.Command {
@@ -33,7 +35,14 @@ func NewCommand(globalFlags *flags.GlobalFlags) *cobra.Command {
 
 	authCmd.Flags().IntVar(&cmd.Port, "port", 0, "port on which the auth server will listen (default 0)")
 	authCmd.Flags().StringVar(&cmd.LocalBindAddress, "local-bind-address", "127.0.0.1", "local address on which the auth server will listen")
-	authCmd.Flags().StringVar(&cmd.RedirectURLHostname, "redirect-url-hostname", "localhost", "hostname of the redirect URL")
+
+	// Declare the version flag and then you can deprecate it.
+	authCmd.Flags().StringVar(&cmd.RedirectURLHostname, "redirect-url-hostname", "", "hostname of the redirect URL")
+	err := authCmd.Flags().MarkDeprecated("redirect-url-hostname", "use --redirect-url instead")
+	if err != nil {
+		panic(fmt.Sprintf("error marking flag --redirect-url-hostname as deprecated: %v", err))
+	}
+	authCmd.Flags().StringVar(&cmd.RedirectURL, "redirect-url", "", "URL of the redirect URL to use for authentication, (e.g. http://localhost:12345)")
 
 	return authCmd
 }
@@ -48,12 +57,43 @@ func (cmd *AuthCmd) Run(cobraCmd *cobra.Command, args []string) error {
 		_ = cli.Stop()
 	}()
 
+	// TODO: Remove this check in the v6 release.
+	// --redirect-url-hostname is deprecated, so we keep it for backward compatibility.
+	if cmd.RedirectURLHostname != "" && cmd.Port == 0 {
+		return fmt.Errorf("--port is required when using --redirect-url-hostname")
+	}
+	if cmd.RedirectURL != "" && cmd.RedirectURLHostname != "" {
+		return fmt.Errorf("--redirect-url and --redirect-url-hostname cannot be used together")
+	}
+	// End of TODO
+
+	// Validate the redirect URL if it is set.
+	if cmd.RedirectURL != "" {
+		_, err := url.ParseRequestURI(cmd.RedirectURL)
+		if err != nil {
+			return fmt.Errorf("invalid redirect URL '%s'", cmd.RedirectURL)
+		}
+	}
+
+	// If redirect URL is set, we require the port to be specified as well.
+	if cmd.RedirectURL != "" && cmd.Port == 0 {
+		return fmt.Errorf("--port is required when using --redirect-url")
+	}
+
 	// customize authentication options based on the command line parameters
 	authOptions := app.AuthenticationOptions{}
 	authOptions.LocalServerBindAddress = fmt.Sprintf("%s:%d", cmd.LocalBindAddress, cmd.Port)
 
+	// TODO: Remove this check in the v6 release.
+	// If redirect URL hostname is set, we use it to construct the redirect URL.
 	if cmd.RedirectURLHostname != "" {
-		authOptions.RedirectURLHostname = cmd.RedirectURLHostname
+		authOptions.RedirectURL = fmt.Sprintf("http://%s:%d", cmd.RedirectURLHostname, cmd.Port)
+	}
+	// End of TODO
+
+	// If redirect URL is set, we use it to construct the redirect URL.
+	if cmd.RedirectURL != "" {
+		authOptions.RedirectURL = cmd.RedirectURL
 	}
 
 	_, err = cli.AuthenticateFromWeb(ctx, authOptions)